+++ src/KeyVault/KeyVault/Commands/AddAzureKeyVaultKey.cs
@@ -232,6 +232,10 @@ namespace Microsoft.Azure.Commands.KeyVault
             if (ShouldProcess(Name, Properties.Resources.AddKey))
             {
                 PSKeyVaultKey keyBundle;
+                if (InputObject != null)
+                {
+                    VaultName = InputObject.VaultName.ToString();
+                }
 
                 if (string.IsNullOrEmpty(KeyFilePath))
                 {
@@ -239,8 +243,7 @@ namespace Microsoft.Azure.Commands.KeyVault
                             VaultName,
                             Name,
                             CreateKeyAttributes(),
-                            Size,
-                            null);
+                            Size);
                 }
                 else
                 {