+++ src/KeyVault/KeyVault/Commands/GetAzureKeyVault.cs
@@ -102,10 +102,11 @@ namespace Microsoft.Azure.Commands.KeyVault
             {
                 case GetVaultParameterSet:
                     ResourceGroupName = string.IsNullOrWhiteSpace(ResourceGroupName) ? GetResourceGroupName(VaultName) : ResourceGroupName;
+                    PSKeyVault vault = null;
 
                     if (ShouldGetByName(ResourceGroupName, VaultName))
                     {
-                        PSKeyVault vault = KeyVaultManagementClient.GetVault(
+                        vault = KeyVaultManagementClient.GetVault(
                                                     VaultName,
                                                     ResourceGroupName,
                                                     ActiveDirectoryClient);