+++ src/KeyVault/KeyVault/Models/ByokWebKeyConverter.cs
@@ -12,13 +12,11 @@
 // limitations under the License.
 // ----------------------------------------------------------------------------------
 
+using Microsoft.Azure.KeyVault.WebKey;
 using System;
 using System.IO;
 using System.Security;
 using KeyVaultProperties = Microsoft.Azure.Commands.KeyVault.Properties;
-using Track2Sdk = Azure.Security.KeyVault.Keys;
-using Track1Sdk = Microsoft.Azure.KeyVault.WebKey;
-using System.Security.Cryptography;
 
 namespace Microsoft.Azure.Commands.KeyVault.Models
 {
@@ -32,22 +30,12 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
             this.next = next;
         }
 
-        public Track1Sdk.JsonWebKey ConvertKeyFromFile(FileInfo fileInfo, SecureString password)
+        public JsonWebKey ConvertKeyFromFile(FileInfo fileInfo, SecureString password)
         {
             if (CanProcess(fileInfo))
                 return Convert(fileInfo.FullName);
             else if (next != null)
                 return next.ConvertKeyFromFile(fileInfo, password);
-            else
-                throw new ArgumentException(string.Format(KeyVaultProperties.Resources.UnsupportedFileFormat, fileInfo.Name));
-        }
-
-        public Track2Sdk.JsonWebKey ConvertToTrack2SdkKeyFromFile(FileInfo fileInfo, SecureString password)
-        {
-            if (CanProcess(fileInfo))
-                return ConvertToTrack2SdkJsonWebKey(fileInfo.FullName);
-            else if (next != null)
-                return next.ConvertToTrack2SdkKeyFromFile(fileInfo, password);
             else
                 throw new ArgumentException(string.Format(KeyVaultProperties.Resources.UnsupportedFileFormat, fileInfo.Name));
         }
@@ -60,31 +48,17 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
             return ByokFileExtension.Equals(fileInfo.Extension, StringComparison.OrdinalIgnoreCase);
         }
 
-        private Track1Sdk.JsonWebKey Convert(string byokFileName)
+        private JsonWebKey Convert(string byokFileName)
         {
             byte[] byokBlob = File.ReadAllBytes(byokFileName);
 
             if (byokBlob == null || byokBlob.Length == 0)
                 throw new ArgumentException(string.Format(KeyVaultProperties.Resources.InvalidKeyBlob, "BYOK"));
-            return new Track1Sdk.JsonWebKey()
+            return new JsonWebKey()
             {
-                Kty = Track1Sdk.JsonWebKeyType.RsaHsm,
+                Kty = JsonWebKeyType.RsaHsm,
                 T = byokBlob,
             };
-        }
-        
-        private Track2Sdk.JsonWebKey ConvertToTrack2SdkJsonWebKey(string byokFileName)
-        {
-            byte[] byokBlob = File.ReadAllBytes(byokFileName);
-
-             if (byokBlob == null || byokBlob.Length == 0)
-                 throw new ArgumentException(string.Format(KeyVaultProperties.Resources.InvalidKeyBlob, "BYOK"));
-
-             return new Track2Sdk.JsonWebKey(new RSACryptoServiceProvider())
-             {
-                 KeyType = Track2Sdk.KeyType.RsaHsm,
-                 T = byokBlob,
-             };
         }
 
         private IWebKeyConverter next;