+++ src/KeyVault/KeyVault/Models/DataServiceCredential.cs
@@ -18,7 +18,6 @@ using System.Linq;
 using System.Threading.Tasks;
 using KeyVaultProperties = Microsoft.Azure.Commands.KeyVault.Properties;
 using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
-using Microsoft.Rest;
 
 namespace Microsoft.Azure.Commands.KeyVault.Models
 {
@@ -70,12 +69,7 @@ namespace Microsoft.Azure.Commands.KeyVault.Models
 
         public string GetToken()
         {
-            return GetAccessToken().AccessToken;
-        }
-
-        public IAccessToken GetAccessToken()
-        {
-            return GetTokenInternal(TenantId, _authenticationFactory, _context, _endpointName).Item1;
+            return GetTokenInternal(this.TenantId, this._authenticationFactory, this._context, this._endpointName).Item1.AccessToken;
         }
 
         private static string GetTenantId(IAzureContext context)